UAF-2494: Moving changes to arizona.edu ProcurementCardDefault
